requirements.txt
streamlit
pandas
docx2txt
pdfplumber
sentence-transformers
faiss-cpu
openai
rapidfuzz
